+++ src/app/app-routing.module.ts
@@ -1,11 +1,33 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import {
+    PathNotFoundComponent,
+    AboutComponent,
+    LoginComponent,
+} from './layout';
 
-const routes: Routes = [];
+const routes: Routes = [
+    {
+        path: 'about',
+        component: AboutComponent,
+    },
+    {
+        path: 'login',
+        component: LoginComponent,
+    },
+    {
+        path: '',
+        redirectTo: '/products-list',
+        pathMatch: 'full',
+    },
+    {
+        path: '**',
+        component: PathNotFoundComponent,
+    },
+];
 
 @NgModule({
     imports: [RouterModule.forRoot(routes)],
     exports: [RouterModule],
 })
-export class AppRoutingModule {
-}
+export class AppRoutingModule {}