+++ src/app/core/models/order.model.ts
@@ -1,0 +1,11 @@
+import { CartProductModel } from './cart-product.model';
+import { OrderStatus } from './order-status.model';
+
+export interface OrderModel {
+    id: string;
+    date: Date;
+    status: OrderStatus;
+    cartProducts: CartProductModel[];
+    totalQuantity: number;
+    total: number;
+}