+++ src/app/orders/components/process-order/process-order.component.ts
@@ -1,0 +1,26 @@
+import { Component, OnInit } from '@angular/core';
+import { OrderService, CartService } from 'src/app/core/services';
+
+@Component({
+    selector: 'app-process-order',
+    templateUrl: './process-order.component.html',
+    styleUrls: ['./process-order.component.scss'],
+})
+export class ProcessOrderComponent implements OnInit {
+    constructor(
+        public orderService: OrderService,
+        public cartService: CartService
+    ) {}
+
+    ngOnInit(): void {}
+
+    onMakeOrder(): void {
+        this.orderService.createOrder(
+            this.cartService.cartProducts,
+            this.cartService.totalQuantity,
+            this.cartService.totalSum
+        );
+        this.cartService.buyAllProducts();
+        alert('Your order has been accepted');
+    }
+}