+++ src/app/orders/orders.module.ts
@@ -1,9 +1,20 @@
 import { NgModule } from '@angular/core';
 
 import { SharedModule } from 'src/app/shared/shared.module';
+import {
+    ProcessOrderComponent,
+    OrderListComponent,
+    OrderItemComponent,
+} from './components';
+import { OrdersRoutingModule } from './orders-routing.module';
 
 @NgModule({
-    declarations: [],
-    imports: [SharedModule],
+    declarations: [
+        ProcessOrderComponent,
+        OrderListComponent,
+        OrderItemComponent,
+    ],
+    imports: [SharedModule, OrdersRoutingModule],
+    exports: [OrderListComponent, OrdersRoutingModule],
 })
 export class OrdersModule {}