+++ src/app/shared/material.module.ts
@@ -6,6 +6,8 @@ import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatListModule } from '@angular/material/list';
 import { MatSelectModule } from '@angular/material/select';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
 
 @NgModule({
     declarations: [],
@@ -18,6 +20,8 @@ import { MatSelectModule } from '@angular/material/select';
         MatButtonModule,
         MatListModule,
         MatSelectModule,
+        MatInputModule,
+        MatFormFieldModule,
     ],
 })
 export class MaterialModule {}